shift-report

+++ minhm80/src/main/java/com/example/minhm80/payload/dto/ShiftReportDTO.java
@@ -27,7 +27,7 @@ public class ShiftReportDTO {
 
     private Double netSale;
 
-    private Double totalOrders;
+    private Integer totalOrders;
 
 
     private UserDto cashier;
